tutorial-typescript/backend/DjangoApi/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
python-decouple
stripe
psycopg2